+++ libs/mops/tests/test_mops/unit/remote/test_adls.py
@@ -1,0 +1,8 @@
+from thds.mops.remote._adls import AdlsFqn, AdlsRoot, to_lazy_uri
+
+
+def test_to_lazy_uri():
+    assert to_lazy_uri(AdlsRoot("thds", "tmp"))() == "adls://thds/tmp/"
+    assert to_lazy_uri(AdlsFqn("thds", "tmp", "foobar/baz"))() == "adls://thds/tmp/foobar/baz"
+    assert to_lazy_uri("adls://thds/tmp/foobar/baz")() == "adls://thds/tmp/foobar/baz"
+    assert to_lazy_uri(lambda: "adls://thds/tmp/foobar/baz")() == "adls://thds/tmp/foobar/baz"